BOCM-AUTOMATIZADO-PT2/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
PyPDF2>=3.0.1
mysql-connector-python>=8.3.0